Changes to Gui
Added Prefix and Message field functions. You can now write MSG# and
itll automatically add it to the prefix- more to come, hopefully.
Clicking on a User in the user list will add it to the prefix and comply
with protocol (ex. MSG#John#).
Holding shift while clicking on a user in the list will add it to the
prefix - this makes it possible to add multiple users.